+++ apps/vanilla/src/scipts/table.ts
